处理5分钟仍未处理完钉钉订单

+++ src/main/java/com/xescm/ofc/service/impl/OfcOrderManageServiceImpl.java
@@ -2793,7 +2793,7 @@ public class OfcOrderManageServiceImpl implements OfcOrderManageService {
             if(null != invalidTime){
                 inventoryDTO.setLotatt02(DateUtils.Date2String(ofcGoodsDetailsInfo.getInvalidTime(), DateUtils.DateFormatType.TYPE2));
             }
-            inventoryDTO.setQuantity(ofcGoodsDetailsInfo.getQuantity().doubleValue());
+            inventoryDTO.setQuantity(ofcGoodsDetailsInfo.getQuantity());
             inventoryGoods.add(inventoryDTO);
         }
         return whcOrderCancelEdasService.validateStockCount(inventoryGoods);
